master bilan merch qildim bazi o'zgarishlar bo'ldi

+++ src/main/java/com/example/entity/Tariff.java
@@ -30,7 +30,7 @@ public class Tariff {
 
     private int branchAmount;
 
-    private int productAmount;
+    private long productAmount;
 
     private int employeeAmount;
